moving ConfigFileOperations from io to config since it depends on the files in config. This should make all dependencies one way: * io and kernel depend only on externals, * config only depends on io, kernel and externals, * batch only depends on io, kernel, config and externals, * gui only depends on io, config, batch and exernals, * main/files (including cli.py) only depends on gui, batch, io and externals
