+++ dfastmi/batch/core.py
@@ -43,9 +43,9 @@ from dfastmi.config.AConfigurationInitializerBase import AConfigurationInitializ
 from dfastmi.config.ConfigurationInitializerFactory import (
     ConfigurationInitializerFactory,
 )
+from dfastmi.config.ConfigFileOperations import ConfigFileOperations
 from dfastmi.io.ApplicationSettingsHelper import ApplicationSettingsHelper
 from dfastmi.io.Branch import Branch
-from dfastmi.io.ConfigFileOperations import ConfigFileOperations
 from dfastmi.io.DFastAnalysisConfigFileParser import DFastAnalysisConfigFileParser
 from dfastmi.io.IReach import IReach
 from dfastmi.io.Reach import Reach
